Change Signedness
Fixes a sign comparison warning!

+++ include/glfwpp/glfwpp.h
@@ -151,7 +151,7 @@ namespace glfw
 
         std::vector<const char*> extensionNames;
         extensionNames.reserve(count);
-        for(int i = 0; i < count; ++i)
+        for(size_t i = 0; i < count; ++i)
         {
             extensionNames.push_back(pExtensionNames[i]);
         }
